pyproject.toml
[project]
name = "aiipython"
version = "0.5.0"
description = "A reactive AI chat assistant running inside IPython"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dspy>=2.6",
    "ipython>=8.0",
    "Pillow>=10.0",
    "rich>=13.0",
    "dspy-template-adapter>=0.2.1",
    "dill>=0.3.9",
    "httpx>=0.27",
]

[project.optional-dependencies]
mlflow = ["mlflow>=3.10"]

[project.scripts]
aiipython = "aiipython.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aiipython"]
include = [
    "src/aiipython/pi_tui_frontend/package.json",
    "src/aiipython/pi_tui_frontend/package-lock.json",
    "src/aiipython/pi_tui_frontend/pi_tui_app.mjs",
    "src/aiipython/pi_gateway/package.json",
    "src/aiipython/pi_gateway/package-lock.json",
    "src/aiipython/pi_gateway/pi_gateway_server.mjs",
    "src/aiipython/pi_native_host/package.json",
    "src/aiipython/pi_native_host/package-lock.json",
    "src/aiipython/pi_native_host/pi_native_host.mjs",
]
